+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 10.15.0
 
+* `[log]` Added support of ANSII colors in the output
 * `[log]` Using `uint8` for level codes instead of `int`
 
 ### 10.14.0